scanspec2: move to scanspec.v2 (Phase 1 package migration)
Nests the 2.0 package inside the existing scanspec distribution
(src/scanspec2/ -> src/scanspec/v2/, tests mirrored) instead of keeping it
as a separate sibling top-level package. 1.x is completely unmodified;
nothing takes over the top-level scanspec name yet.

This is Phase 1 of a two-phase migration (see PRD §12): letting ophyd-async
start integrating against scanspec.v2 immediately, without waiting for or
risking anything on 1.x. Phase 2 (later, separate) promotes v2 to the
top-level name and demotes 1.x to scanspec.v1 once 2.0 is actually ready to
release.

v2 is deliberately not imported from scanspec/__init__.py, staying opt-in
(import scanspec.v2) rather than pulled in by default for 1.x-only
consumers. As a side effect this also keeps it out of the Sphinx docs
build's autosummary recursion for now (verified) -- rewriting docs/ for
the 2.0 API is separate, later work (PRD §12 Phase 2).

Updates AGENTS.md, PRD.md (path references + two-phase §12), and
API_SPEC.md to match. Also fixes an unrelated stale AGENTS.md note
("Known churn") that still described a pre-existing TriggerNode/
trigger_nodes design that was never actually adopted -- the real names
that landed are TriggerRepeat/TriggerSequence and Window.trigger_sequences.

Verified: editable install + import scanspec.v2 works; 219 tests pass on
the new paths; full suite (333 passed, 1 skipped, 1 xfailed) confirms 1.x
is unaffected; pyright and ruff both clean on src/scanspec/v2 and
tests/scanspec/v2.
