+++ AGENTS.md
@@ -5,31 +5,38 @@
 - **`PRD.md`** — requirements and current design intent. Authoritative; read
   it before designing or implementing anything.
 - **`API_SPEC.md`** — annotated consumption-API examples. Partially stale
-  (see PRD §10); where it disagrees with `src/scanspec2/` + PRD, the latter
-  win.
+  (see PRD §10); where it disagrees with `src/scanspec/v2/` + PRD, the
+  latter win.
 - **`docs/explanations/decisions/`** — ADRs. 0001–0005 accepted; 0006
   tentative; 0007 proposed with pending review corrections (PRD §9). Do not
   treat 0006/0007 as settled.
 
 ## Repository structure
 
-Two packages coexist during 2.0 development:
+1.x and 2.0 coexist during 2.0 development, as sibling packages under one
+top-level `scanspec` distribution:
 
-- `src/scanspec/` — the 1.x package. **Do not modify.** Reference only;
-  don't load it into context unless porting a specific algorithm.
-- `src/scanspec2/` — the 2.0 package. All new work goes here.
+- `src/scanspec/` (everything except `v2/`) — the 1.x package. **Do not
+  modify.** Reference only; don't load it into context unless porting a
+  specific algorithm.
+- `src/scanspec/v2/` — the 2.0 package, nested as a submodule
+  (`import scanspec.v2`). All new work goes here. Not yet the top-level
+  `scanspec` name — that happens at final 2.0 release (PRD §12, Phase 2).
 
-Tests mirror this: `tests/` covers 1.x (do not modify); `tests/scanspec2/`
-is where all new tests go.
+Tests mirror this: `tests/` (flat, everything except `scanspec/v2/`) covers
+1.x (do not modify); `tests/scanspec/v2/` is where all new tests go.
 
 Branch flow: feature branches → PRs against `bluesky/scanspec:v2-dev` →
 `v2-dev` merges to `main` only at the final 2.0 migration (PRD §12).
 
 ## Known churn — check before building on these
 
-- `TriggerPattern`/`TriggerGroup` will merge into a `TriggerNode` tree and
-  `Window.trigger_groups` → `Window.trigger_nodes` when ADR 0007 is
-  accepted. Avoid new code that deepens coupling to the current split.
+- `Acquire.detectors: Sequence[DetectorGroup]` may be replaced —
+  `_bake_trigger_sequence` currently guesses the parent/child hierarchy by
+  ranking `DetectorGroup`s by duration; the maintainer wants `Acquire` to
+  stop making that decision, with the caller supplying an already-built
+  hierarchy instead. Not yet designed (PRD §11). Avoid new code that
+  deepens coupling to the current auto-ranking logic.
 - Naming that the docs sometimes get wrong: the code uses
   `Window.non_linear` (not `non_linear_move`), `Scan.has_moving_axes` /
   `Scan.non_linear` (there is no `Scan.fly`), and
@@ -41,7 +48,7 @@ Branch flow: feature branches → PRs against `bluesky/scanspec:v2-dev` →
 - Simple, direct assertions; test **public interfaces**; avoid mocks unless
   there is no other way.
 - No serialisation tests for plain dataclasses (they carry none).
-- **`tests/scanspec2/test_use_cases.py` is the maintainer's file.** Never
+- **`tests/scanspec/v2/test_use_cases.py` is the maintainer's file.** Never
   add, remove, or modify tests in it without explicit permission. Put your
   tests in `test_compile.py`, `test_core.py`, `test_specs.py`, etc.
 - **Assert real, independently-derived expected values — not just shape,
@@ -60,9 +67,9 @@ Branch flow: feature branches → PRs against `bluesky/scanspec:v2-dev` →
 ## Quality gates — all three must pass after every change
 
 ```bash
-pytest tests/scanspec2/ -v
-python -m pyright src/scanspec2/ tests/scanspec2/   # 0 errors
-ruff check src/scanspec2/ tests/scanspec2/          # 0 errors
+pytest tests/scanspec/v2/ -v
+python -m pyright src/scanspec/v2/ tests/scanspec/v2/   # 0 errors
+ruff check src/scanspec/v2/ tests/scanspec/v2/          # 0 errors
 ```
 
 ## Type annotations and lint
