+++ src/scanspec/v2/__init__.py
@@ -1,5 +1,5 @@
 """scanspec 2.0 — work in progress."""
 
-from scanspec2.core import TriggerRepeat, TriggerSequence
+from .core import TriggerRepeat, TriggerSequence
 
 __all__ = ["TriggerRepeat", "TriggerSequence"]
