+++ src/scanspec/v2/core.py
