+++ src/scanspec/v2/specs.py
