+++ tests/scanspec/v2/test_specs.py
@@ -1,4 +1,4 @@
-"""Tests for scanspec2.specs — motion nodes, operator algebra, Acquire validation."""
+"""Tests for scanspec.v2.specs — motion nodes, operator algebra, Acquire validation."""
 
 from typing import Any
 
@@ -6,8 +6,8 @@ import numpy as np
 import pytest
 from pydantic import TypeAdapter, ValidationError
 
-from scanspec2.core import ContinuousStream, DetectorGroup, MonitorStream
-from scanspec2.specs import (
+from scanspec.v2.core import ContinuousStream, DetectorGroup, MonitorStream
+from scanspec.v2.specs import (
     Acquire,
     AnySpec,
     Concat,
@@ -304,7 +304,7 @@ def test_linspace_bounded_symmetric():
 
 
 def test_range_positional():
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     r = Range("x", 0.0, 1.0, 0.25)
     assert r.axis == "x"
@@ -314,28 +314,28 @@ def test_range_positional():
 
 
 def test_range_keyword():
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     r = Range(axis="x", start=0.0, stop=10.0, step=2.0)
     assert r.step == 2.0
 
 
 def test_range_zero_step_raises():
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     with pytest.raises(ValueError):
         Range("x", 0.0, 1.0, 0.0)
 
 
 def test_range_negative_step_raises():
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     with pytest.raises(ValueError):
         Range("x", 0.0, 1.0, -0.5)
 
 
 def test_range_type_field():
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     r = Range("x", 0.0, 1.0, 0.5)
     assert r.type == "Range"
@@ -347,7 +347,7 @@ def test_range_type_field():
 
 
 def test_range_bounded_many_points():
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     inst = Range.bounded("x", 0.0, 1.0, 0.25)
     assert isinstance(inst, Range)
@@ -367,7 +367,7 @@ def test_range_bounded_many_points():
 def test_range_bounded_one_point(
     lower: float, upper: float, step: float, expected_start: float
 ) -> None:
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     inst = Range.bounded("x", lower, upper, step)
     assert isinstance(inst, Range)
@@ -376,7 +376,7 @@ def test_range_bounded_one_point(
 
 def test_range_bounded_lower_equals_upper():
     """lower == upper must not crash and must produce a single point."""
-    from scanspec2.specs import Range
+    from scanspec.v2.specs import Range
 
     inst = Range.bounded("x", 5.0, 5.0, 0.5)
     assert isinstance(inst, Range)
@@ -386,13 +386,13 @@ def test_range_bounded_lower_equals_upper():
 
 
 def test_line_is_linspace():
-    from scanspec2.specs import Line
+    from scanspec.v2.specs import Line
 
     assert Line is Linspace
 
 
 def test_line_instantiation():
-    from scanspec2.specs import Line
+    from scanspec.v2.specs import Line
 
     ln = Line("x", 0.0, 10.0, 5)
     assert isinstance(ln, Linspace)
@@ -406,7 +406,7 @@ def test_line_instantiation():
 
 
 def test_spiral_positional():
-    from scanspec2.specs import Spiral
+    from scanspec.v2.specs import Spiral
 
     s = Spiral("x", 0.0, 5.0, 2.0, "y", 10.0, 10.0)
     assert s.x_diameter == 5.0
@@ -415,7 +415,7 @@ def test_spiral_positional():
 
 
 def test_spiral_y_diameter_defaults_to_x_diameter():
-    from scanspec2.specs import Spiral
+    from scanspec.v2.specs import Spiral
 
     s_implicit = Spiral("x", 0.0, 5.0, 2.0, "y", 10.0)
     s_explicit = Spiral("x", 0.0, 5.0, 2.0, "y", 10.0, y_diameter=5.0)
@@ -428,7 +428,7 @@ def test_spiral_y_diameter_defaults_to_x_diameter():
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_spiral_x_diameter_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Spiral
+    from scanspec.v2.specs import Spiral
 
     with pytest.raises(ValueError):
         Spiral("x", 0.0, bad_value, 2.0, "y", 10.0)
@@ -436,7 +436,7 @@ def test_spiral_x_diameter_not_positive_raises(bad_value: float):
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_spiral_x_step_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Spiral
+    from scanspec.v2.specs import Spiral
 
     with pytest.raises(ValueError):
         Spiral("x", 0.0, 5.0, bad_value, "y", 10.0)
@@ -444,7 +444,7 @@ def test_spiral_x_step_not_positive_raises(bad_value: float):
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_spiral_y_diameter_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Spiral
+    from scanspec.v2.specs import Spiral
 
     with pytest.raises(ValueError):
         Spiral("x", 0.0, 5.0, 2.0, "y", 10.0, y_diameter=bad_value)
@@ -456,7 +456,7 @@ def test_spiral_y_diameter_not_positive_raises(bad_value: float):
 
 
 def test_ellipse_positional():
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     e = Ellipse("x", 5.0, 1.0, 0.5, "y", 0.0)
     assert e.x_axis == "x"
@@ -468,7 +468,7 @@ def test_ellipse_positional():
 
 
 def test_ellipse_y_diameter_defaults_to_x_diameter():
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     e_implicit = Ellipse("x", 0.0, 2.0, 0.5, "y", 0.0)
     e_explicit = Ellipse("x", 0.0, 2.0, 0.5, "y", 0.0, y_diameter=2.0)
@@ -480,7 +480,7 @@ def test_ellipse_y_diameter_defaults_to_x_diameter():
 
 
 def test_ellipse_y_step_defaults_to_x_step():
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     e_implicit = Ellipse("x", 0.0, 2.0, 0.5, "y", 0.0)
     e_explicit = Ellipse("x", 0.0, 2.0, 0.5, "y", 0.0, y_step=0.5)
@@ -492,14 +492,14 @@ def test_ellipse_y_step_defaults_to_x_step():
 
 
 def test_ellipse_vertical_default():
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     e = Ellipse("x", 0.0, 2.0, 0.5, "y", 0.0)
     assert e.vertical is False
 
 
 def test_ellipse_explicit_y_diameter_and_y_step():
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     e = Ellipse("x", 0.0, 4.0, 1.0, "y", 0.0, y_diameter=2.0, y_step=0.5)
     assert e.y_diameter == 2.0
@@ -508,7 +508,7 @@ def test_ellipse_explicit_y_diameter_and_y_step():
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_ellipse_x_step_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     with pytest.raises(ValueError):
         Ellipse("x", 0.0, 2.0, bad_value, "y", 0.0)
@@ -516,7 +516,7 @@ def test_ellipse_x_step_not_positive_raises(bad_value: float):
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_ellipse_x_diameter_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     with pytest.raises(ValueError):
         Ellipse("x", 0.0, bad_value, 1.0, "y", 0.0)
@@ -524,7 +524,7 @@ def test_ellipse_x_diameter_not_positive_raises(bad_value: float):
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_ellipse_y_diameter_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     with pytest.raises(ValueError):
         Ellipse("x", 0.0, 2.0, 1.0, "y", 0.0, y_diameter=bad_value)
@@ -532,7 +532,7 @@ def test_ellipse_y_diameter_not_positive_raises(bad_value: float):
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_ellipse_y_step_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Ellipse
+    from scanspec.v2.specs import Ellipse
 
     with pytest.raises(ValueError):
         Ellipse("x", 0.0, 2.0, 1.0, "y", 0.0, y_step=bad_value)
@@ -544,7 +544,7 @@ def test_ellipse_y_step_not_positive_raises(bad_value: float):
 
 
 def test_polygon_positional():
-    from scanspec2.specs import Polygon
+    from scanspec.v2.specs import Polygon
 
     vertices = [(0.0, 0.0), (5.0, 0.0), (2.5, 4.0)]
     p = Polygon("x", "y", vertices, 1.0)
@@ -555,7 +555,7 @@ def test_polygon_positional():
 
 
 def test_polygon_y_step_defaults_to_x_step():
-    from scanspec2.specs import Polygon
+    from scanspec.v2.specs import Polygon
 
     p_implicit = Polygon("x", "y", [(0.0, 0.0), (1.0, 0.0), (0.5, 1.0)], 0.25)
     p_explicit = Polygon(
@@ -569,7 +569,7 @@ def test_polygon_y_step_defaults_to_x_step():
 
 
 def test_polygon_explicit_y_step():
-    from scanspec2.specs import Polygon
+    from scanspec.v2.specs import Polygon
 
     p = Polygon("x", "y", [(0.0, 0.0), (1.0, 0.0), (0.5, 1.0)], 0.5, 0.25)
     assert p.x_step == 0.5
@@ -577,7 +577,7 @@ def test_polygon_explicit_y_step():
 
 
 def test_polygon_vertical_default():
-    from scanspec2.specs import Polygon
+    from scanspec.v2.specs import Polygon
 
     p = Polygon("x", "y", [(0.0, 0.0), (1.0, 0.0), (0.5, 1.0)], 0.25)
     assert p.vertical is False
@@ -585,7 +585,7 @@ def test_polygon_vertical_default():
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_polygon_x_step_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Polygon
+    from scanspec.v2.specs import Polygon
 
     with pytest.raises(ValueError):
         Polygon("x", "y", [(0.0, 0.0), (1.0, 0.0), (0.5, 1.0)], bad_value)
@@ -593,7 +593,7 @@ def test_polygon_x_step_not_positive_raises(bad_value: float):
 
 @pytest.mark.parametrize("bad_value", [0.0, -1.0])
 def test_polygon_y_step_not_positive_raises(bad_value: float):
-    from scanspec2.specs import Polygon
+    from scanspec.v2.specs import Polygon
 
     with pytest.raises(ValueError):
         Polygon("x", "y", [(0.0, 0.0), (1.0, 0.0), (0.5, 1.0)], 1.0, y_step=bad_value)
